Refactor impersonation constraint from org:Role to rdfs:Resource
We assessed the compatibility of the original service with lblod-projects.
We concluded that an account is a bit more complicated as in kaleidos, as it uses roles and groups (in most cases at least).
We also deduced that maybe all kinds of resources could be impersonated.
That's why we abstracted `org:Role` to `rdfs:Resource` as the target for impersonated sessions.
We think mu-auth should enforce which specific types of resources can be impersonated.
We think the client should figure out (or assume) which specific type
of resource is being impersonated.

We also moved away from
`<http://mu.semte.ch/vocabularies/ext/impersonatedRole>` to
`<http://mu.semte.ch/vocabularies/account/impersonation/impersonates>`