Fix the setImpersonatedSession query
The switch flow wasn't working properly. When switching to a different account impersonation the roles and groups would be copied to the `ext:originalX` predicates because the FILTER NOT EXISTS part of the query was wrong. If you then stopped impersonating the roles would still be attached to the admin account.